chart.vueに専念
